Clean up/refactor
