--- src/pages/configurePatch/configurePatchCore/state.ts
@@ -1,9 +1,0 @@
-export type DownstreamPatchState = {
-  [alias: string]: boolean;
-};
-export type TasksState = {
-  [task: string]: boolean;
-};
-export type VariantTasksState = {
-  [variant: string]: TasksState;
-};
